incorrect, but fully working program
The problem is determine where is the silence lavel. In this code I find out the peaks and think that they are real notes. But they aren't real notes. May be it wil be reasonable to find out the maximum for each block and looking up the notes among "delta down" (delta_peak) for this maximum