fix another stupid mistake
I wasn't checking the types that was comment before (but I deleted this comment without checking - that's full mistake!). So, now all is Ok. To add, I change a policy with tirning frequences into notes in function freqToNote: now I choose the nearest frequance instead of the frequance with maximum amplitude.